% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Answer.R
\name{Answer}
\alias{Answer}
\title{R6 class representing an answer}
\description{
This class represents an answer.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{list}}{returns the answer as a list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Answer-new}{\code{Answer$new()}}
\item \href{#method-Answer-print}{\code{Answer$print()}}
\item \href{#method-Answer-clone}{\code{Answer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Answer-new"></a>}}
\if{latex}{\out{\hypertarget{method-Answer-new}{}}}
\subsection{Method \code{new()}}{
This function initializes the answer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Answer$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a list containing the data of the answer}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The answer object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Answer-print"></a>}}
\if{latex}{\out{\hypertarget{method-Answer-print}{}}}
\subsection{Method \code{print()}}{
this function prints the answer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Answer$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Answer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Answer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Answer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
