% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIFTParser.R
\name{GIFTParser}
\alias{GIFTParser}
\title{Parse a GIFT formatted text into a list of questions}
\usage{
GIFTParser(text, debug = FALSE)
}
\arguments{
\item{text}{A character vector containing the GIFT formatted text to parse}

\item{debug}{A logical value indicating if the parser should be run in debug mode}
}
\value{
A list of questions
}
\description{
Parse a GIFT formatted text into a list of questions
}
