% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Question.R
\name{Question}
\alias{Question}
\title{R6 class representing a question}
\description{
This class represents a question. It has a title, a text, a category and a type.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Question-new}{\code{Question$new()}}
\item \href{#method-Question-build_answers}{\code{Question$build_answers()}}
\item \href{#method-Question-export}{\code{Question$export()}}
\item \href{#method-Question-import}{\code{Question$import()}}
\item \href{#method-Question-clone}{\code{Question$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-new"></a>}}
\if{latex}{\out{\hypertarget{method-Question-new}{}}}
\subsection{Method \code{new()}}{
This function initializes the question with a title
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list containing the data of the question}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The question object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-build_answers"></a>}}
\if{latex}{\out{\hypertarget{method-Question-build_answers}{}}}
\subsection{Method \code{build_answers()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$build_answers(answers)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-export"></a>}}
\if{latex}{\out{\hypertarget{method-Question-export}{}}}
\subsection{Method \code{export()}}{
This function returns the question as a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$export()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the question
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-import"></a>}}
\if{latex}{\out{\hypertarget{method-Question-import}{}}}
\subsection{Method \code{import()}}{
This function imports a question from a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$import(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list containing the question data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The question object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Question-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
