% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Question.R
\name{Question}
\alias{Question}
\title{R6 class representing a question}
\description{
This class represents a question. It has a title, a text, a category and a type.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{list}}{returns the question as a list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Question-new}{\code{Question$new()}}
\item \href{#method-Question-import_gift}{\code{Question$import_gift()}}
\item \href{#method-Question-export_gift}{\code{Question$export_gift()}}
\item \href{#method-Question-import}{\code{Question$import()}}
\item \href{#method-Question-print}{\code{Question$print()}}
\item \href{#method-Question-clone}{\code{Question$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-new"></a>}}
\if{latex}{\out{\hypertarget{method-Question-new}{}}}
\subsection{Method \code{new()}}{
This function initializes the question with a title
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list containing the data of the question}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The question object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-import_gift"></a>}}
\if{latex}{\out{\hypertarget{method-Question-import_gift}{}}}
\subsection{Method \code{import_gift()}}{
This function imports a question from a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$import_gift(string)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{string}}{A string containing the question data in GIFT format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The question object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-export_gift"></a>}}
\if{latex}{\out{\hypertarget{method-Question-export_gift}{}}}
\subsection{Method \code{export_gift()}}{
This function returns the question as a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$export_gift()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list representation of the question
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-import"></a>}}
\if{latex}{\out{\hypertarget{method-Question-import}{}}}
\subsection{Method \code{import()}}{
This function imports a question from a list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$import(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list containing the question data}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The question object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-print"></a>}}
\if{latex}{\out{\hypertarget{method-Question-print}{}}}
\subsection{Method \code{print()}}{
This function prints the question
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Question-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Question-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Question$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
