% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_Quiz.R
\name{Quiz}
\alias{Quiz}
\title{R6 Class representing a Quiz}
\description{
This class represents a Quiz object. It is a container for a list of questions.
}
\details{
The Quiz class is a container for a list of questions. It is used to store
and manipulate questions.
Questions can be added to the Quiz object using the add_question method, or
directly using a GIFT file with the import_gift method.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{list}}{This function returns the list of questions}

\item{\code{description}}{This function returns the description of the Quiz object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Quiz-new}{\code{Quiz$new()}}
\item \href{#method-Quiz-add_question}{\code{Quiz$add_question()}}
\item \href{#method-Quiz-import_gift}{\code{Quiz$import_gift()}}
\item \href{#method-Quiz-clone}{\code{Quiz$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-new"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-new}{}}}
\subsection{Method \code{new()}}{
This function initializes the Quiz object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Quiz object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-add_question"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-add_question}{}}}
\subsection{Method \code{add_question()}}{
This function adds a question to the Quiz object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$add_question(question)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{question}}{The question to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-import_gift"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-import_gift}{}}}
\subsection{Method \code{import_gift()}}{
This function imports a GIFT formatted text into the Quiz object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$import_gift(text)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{A character vector containing the GIFT formatted text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quiz-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Quiz-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quiz$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
